Move PublishingMixin into sprockets_amqp.web.
