+++ docs/api.rst
@@ -3,7 +3,14 @@ API
 .. automodule:: sprockets.mixins.amqp
    :members:
 
+AMQP Client
+-----------
 .. automodule:: sprockets_amqp.client
+   :members:
+
+Tornado Web Integration
+-----------------------
+.. automodule:: sprockets_amqp.web
    :members:
 
 Exceptions
