+++ sprockets/mixins/amqp/__init__.py
@@ -16,20 +16,21 @@ The ``AMQP``` prefix is interchangeable with ``RABBITMQ``. For example, you can
 use ``AMQP_URL`` or ``RABBITMQ_URL``.
 
 """
-import os
 import logging
+import os
 import sys
 
 import sprockets_amqp.exceptions  # can be safely imported
 
 try:
     from sprockets_amqp.client import Client
+    from sprockets_amqp.web import PublishingMixin
     from tornado import concurrent, ioloop
     import pika.exceptions
 except ImportError:  # pragma: nocover
     sys.stderr.write('setup.py import error compatibility objects created\n')
-    concurrent, ioloop, pika, Client = \
-        object(), object(), object(), object()
+    concurrent, ioloop, pika, Client, PublishingMixin = \
+        object(), object(), object(), object(), object()
 
 
 LOGGER = logging.getLogger(__name__)
@@ -104,31 +105,3 @@ def install(application, io_loop=None, **kwargs):
     LOGGER.debug('kwargs: %r', kwargs)
     setattr(application, 'amqp', Client(**kwargs))
     return True
-
-
-class PublishingMixin(object):
-    """This mixin adds publishing messages to RabbitMQ. It uses a
-    persistent connection and channel opened when the application
-    start up and automatically reopened if closed by RabbitMQ
-
-    """
-
-    def amqp_publish(self, exchange, routing_key, body, properties=None):
-        """Publish a message to RabbitMQ
-
-        :param str exchange: The exchange to publish the message to
-        :param str routing_key: The routing key to publish the message with
-        :param bytes body: The message body to send
-        :param dict properties: An optional dict of AMQP properties
-        :rtype: tornado.concurrent.Future
-
-        :raises: :exc:`sprockets_amqp.exceptions.AMQPException`
-        :raises: :exc:`sprockets_amqp.exceptions.NotReadyError`
-        :raises: :exc:`sprockets_amqp.exceptions.PublishingFailure`
-
-        """
-        properties = properties or {}
-        if hasattr(self, 'correlation_id') and getattr(self, 'correlation_id'):
-            properties.setdefault('correlation_id', self.correlation_id)
-        return self.application.amqp.publish(
-            exchange, routing_key, body, properties)
